backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-limiter==3.5.0
bcrypt==4.1.2
python-dotenv==1.0.0
google-generativeai==0.3.2
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
python-multipart==0.0.6
pillow==10.1.0
email-validator==2.1.0
pydantic==2.5.2 